requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
rembg==2.0.67
onnxruntime==1.22.1
pillow==11.3.0
boto3==1.39.9
redis==5.2.1
rq==1.16.2
pytest==8.4.1
